Improve role distribution
It often happen that some nodes aren't deployed correctly.
When it happens it's maybe better to not use them in the subsequent
operations (ssh connections).
So roles are distributed from the deployed node.
When using a single cluster we garantee that

* each role has at least one node (allowing OS to run)
* one node belongs to at most one role

NB: With this choice it can happen that some roles doesn't have all the
desired nodes (in our case, some compute nodes may be missing).