Add support for dynamic roles.
This patch is guided by the fact that we'll have to deploy
Openstack using different topology (some services on dedicated nodes,
some services grouped together ...).

Until now changing the topology affects many files including python
code like

- `reservation.yaml` to add the new role
- `templates/multinode.jinja2` to declare this role to ansible
- `templates/multinode.jinja2` to modify the super groups to which this
  role belongs (e.g run mariadb on a dedicated node - see below)
- `01_deploy_nodes.py` to add the node to the template vars.
-  ansible code in case of the new service added

The proposed patch show that this is possible to achieve this kind of
modification without changing the python code.

Moreover it's done without *polluting* the git repository.
Indeed `templates/multinode.jinja2` is a versionned file, any change to
it will be tracked. As a consequence two users testing various
deployment topologies will see their git diverged.  Instead this patch let
the user define its own inventory without being added to the git
history. Additionnaly when a consensus is found on an alternative inventory
it can be versionned as a possible alternative

What's new ?

 - files have been reorganized:
    - inventories directory contains the inventory you can use
    - ansible directory contains the ansible code for the first playbook
      to run
 - resources roles must be named after kolla terminology : control,
   storage, compute, network
 - inventory file is generated in current/multinode and is the
   concatenation of the generated roles and the static part that is
   found under the inventories directory
 - enable_monitoring option: enable/disable monitoring stack
 - inventory option : points to the inventory to use.
 - inventories directory isn't tracked
 - reservation.yaml contains kolla/globals variable : there is no need
   to use a template engine for a yaml file. It centralized to
   configuration of the experimentation.
   The actual globals.yaml is then generated in current/globals.yml
 - influx is virtualized : that means it can run wherever you wants
   dedicated node/ grouped with other ...
 - grafana is virtualized as well
 - cadvisor now reports the machine metric in the format
   role-machine_name

Let's take an example.
I want to run mariadb on a dedicated node :

1) `reservation.yaml`

resources:
    cluster:
        control     : x
        storage     : x
        mariadb-node: 1

By adding this the groupe mariadb-node will be added to the inventory
dynamically and will contains the node reserved and deploy on g5k

Note: mariadb is a reserved name in the original kolla inventory

2) create a new inventory by copy pasting the sample given
Update the mariadb group :

```
[mariadb:children]
mariadb-node
```

3) Point your inventory to the newly created one in `reservation.yaml`