Allow custom dir with default to the symlink dir
backup_dir is a variable that is passed in the environnment before
backuping any data.
This directory is created if it does not exist.
This directory can be specified through the command line :
```
python -m enos.enos backup --backup_dir=BACKUP_DIR
```
if BACKUP_DIR isn't an absolute path it is considered relative to the
current working directory (directory where the command is called)
if BACKUP_DIR is not specified it is defaulted to the SYMLINK_NAME.

fix #17