Network emulation (latency and bandwidth constraints)
Groups are introduced in the configuration file in order to easily describe
network constraints between them.  Constraints are enforced using a new
task : `tc`.  From a high level point of view this command will apply a set of
traffic shapping rules according to the group description.

Groups are described using the configuration file :
```
topology:
  grp1:
    parasilo:
      control: 1
      network: 1
  grp[2-4]:
    parasilo:
      compute: 1
```

Note that group expansion are possible, `grp[2-4]` is a concise way to
describe grp2, grp3 and grp4
Note that old `resources` is still supported. But specifying a `topology key
will override it.

Network constraints are described through the following matrix
(excerpt) in the configuration file:

```
network_constraints:
  default_delay: 50ms
  default_rate: 1gbit
  constraints:
    - src: grp1
      dst: grp[2-4]
      delay: 100ms
      symetric: true
```

Note that group expansion is also possible in `network_constraints`.
Note that `symetric` is optionnal (false by default).
Note that an `enable` key is also available. Setting it to `false` will
reset all the rules enforced.

Workflow:

* At deployment time, groups are merged to create the `resources` description as
it would have been previously (without groups)

* During the provider init, nodes are distributed into roles (as before) and into
groups (new feature). More precisely the `env[rsc]` will contains roles (as
before) and groups (new). As a consequence the generated `multinode` file will
also contain the groups descriptions.

* After the OpenStack deployment the `tc` command will :

1) Retrieve all the ips of all the nodes in play for all interfaces.
This is done through a dedicated ansible playbook : `utils/ips.yml`

2) Build the description of the rules to apply on nodes.
A rule is composed of a source node, a device (on which to apply the rule), the
destination ip, the desired rate and delay. For this particular step, we thus
need to calculate the cartesian product of the informations given by the first
step.

3) Apply the rules. This is done with an ansible playbool : `utils/tc.yml`

Additionnaly a quick check can be performed using the `utils/test.yml` playbook.

Further Notes :
 * Groups (grp1, grp2 ...) in the `multinode` are unused for now.
 * Enforcing network constraints is slow in this implementation as every
node receive all the rules. Alternatively we could send to the nodes its own
rules.
 * At the time of the implementation, I chose to put the network
constraints in the environment. It can makes the env much bigger than
before.