Upgrade Enos for debian/jessie
The goal is to rely as much as possible on official supported images.

provider) python-pip is broken so we use easy-install to get pip.

ceph) the mapping was hard coded and worked for the ubuntu we
deployed but was problematic with jessie.

systemd ) jessie is using systemd, updated where necessary