Rewrite build_roles function
Making it more concise, clear (hopefully) and agnostic
to the provider used.
The role of this function is to distribute actual nodes to the
different roles specified in the configuration file. It has to be
reiterable in the sense that subsequent calls must lead to the same
distribution of nodes. Additionnaly it must handle the case where
less actual nodes are available than requested.

It takes any dict representing a resource description (could
be a "topology" or a regular "resources" description or a dict of
any depth) :

```
resources:
    large:
	control:1
	compute:2
    medium:
	compute:2
	network:1
    tiny:
	compute:2
```
It then builds
* all the resources claims in the form of a list of list (rindexes) :
* the pools of node available grouped by a key function (cluster
name for G5k, size for vbox)

```
[[large, control, 1], [large, compute, 2], [medium, compute, 2],
[medium, network, 1], [tiny, compute, 2]]
```

```
{
    large: [n1, n2, n3],
    medium: [n4, n5, n6],
    tiny: [n7, n8]
}
```

Finally it iterates over all element of rindexes, and takes nodes
from the corresponding pool.