G5K provider: support multiple queues in resa
In cases where the reservation.yaml includes nodes from different
queues (which is typically the case when a new cluster is still
in the 'testing' queue, or when using nodes from the 'default'
queue along with nodes from the 'production' queue.

The `QUEUE_PRIORITIES` indicate which queue to use for nodes that
have multiple queues (actually the general case, as all nodes are
in the 'admin' queue in addition do the 'default' or 'production'
queue. 'admin' is not present in that list to avoid using it, but
it can be added. The default list gives priority to the 'default'
queue over the 'production' queue.

When multiple queues are involved, one oargrid job per queue is
created. Then everything operates as before, the provider simply
merges resources form the jobs.

The only notable differences are that:
- some output have changed to display all the jobs
- job names are now suffixed with an integer counter. So by default
if the job name is 'enos', the created job will be named 'enos-0'.