Specify the provider into the configuration file. (#87)
Resolves #79

Removes the support of `--provider=<>` from the command line and puts
this information into the reservation.yml. The user should specify a
provider name either with a String (ie, `provider: "g5k"`) or under a
specific key called type (ie, `provider: type: "g5k"`). The second
specification lets the provider add extra key under `provider:` (See
msimonin comment on #79).

About the implementation, I remove the building of the provider from
the enostask and move it into phases that need it, because keeping the
building in enostask makes the implementation really tricky.