Add Debian unfree apt repository (for netperf)
Netperf is available into the unfree directory. This commit add the
unfree repository to apt/source.list, so we can install netperf. Note
that the unfree repository is activated by default in G5k debian9 kaenv.

For unbuntu, netperf is available into multiverse repo. The multiverse
repo seems activated by default. At least, this is the case for the
vagrant box ubuntu/xenial64.