Fix kolla command
There was an implicit assumption when using the kolla command : `enos
kolla` can be only run after `enos os`. Indeed the kolla source
directory needed be installed in the experiment directory.

In order to be able to use the `kolla` command even if the `os` command
wasn't run before, we check if the sources are present and download them
if necessary. This code is shared between the install_os and kolla
command.

Note that `enos up` must be called before in any case.

So now the behaviour is

- enos os: install and bootstrap kolla sources and overwrite them if a previous installation is
    present

enos kolla: installs and bootstrap kolla sources only if a previous installation isn't
    present