[EnOSlib][EnOS]Handle the single interface case
When a provider doesn't assign the role neutron_network_interface
the framework provision a dummy interface associated with this role.
In the init phase, external ips will be taken in the other
networks( associated to the network_interface, or api_interface role).
Ideally we must take these ips in an extra network regardless the number
of interfaces available on the nodes.

In any case the implementation is provider agnostic.