[G5K] Autodetect the site
Networks reservation requires the site to be specified in EnOSlib. In
the process of switching to EnOSlib we hardcoded the site (rennes). We
know auto-detect the site and thus allow to use EnOS (with EnOSlib
anywhere).

We currently consider two cases:

1) if the deployment is multisite (number of sites > 1) we fail with an
exception. In the future we'll support multisite deployment

2) if the deployment is monosite we set the site in the network
description accordingly.

For #272