Align the code with future enoslib 2.x.y.
Note that this drops py27 support. This patch is minimal in the sense it
only introduce the configuration object from  the existing dictionnary,
in the future we'll probably want to get rid of the intermediate
dictionnary.

What I also see
- ceph registry is currently broken (probably due to the
ceph cluster migration), we'll need to fix that in a future patch
- git need to be put as mandatory package (not installed on
debian9-x64-nfs)