Build the doc with github actions
Change the build of the documentation from readthedocs.io to github actions.
This change is motivated by the inability of readthedocs builder to call
poetry to build the doc.

As a nice side effect, the doc is now pushed to
https://beyondtheclouds.github.io/enos/ which is consistent with the discovery
website.  Unfortunately, we now only build the documentation of the latest
version.  This removes the ability to browse older versions of the
documentation.