up: Fix Docker configuration for kolla-ansible 12
Since kolla-ansible 12, docker is configured to disable container
networking.  But we require networking in order to run our init script in
a kolla-toolbox container, because it needs to talk to the OpenStack API
we just deployed.

Simply disable this new config default to restore regular Docker networking.

This new config default in kolla-ansible was seemingly introduced for
security reasons, but this is not a big concern for Enos (we target
experimental deployments, not production deployments)

Ref: https://opendev.org/openstack/kolla-ansible/commit/4053a0afdb3d0a230557883453b89b06cf4d7057